+++ Source/SelfService/Web/assets/images.d.ts
