+++ Source/SelfService/Web/microservice/components/restartInfoBox.tsx
