+++ Source/SelfService/Web/screens/applicationScreen.tsx
@@ -53,6 +53,6 @@ export const ApplicationScreen: React.FunctionComponent = () => {
                     <Building />
                 </Route>
             </Switch>
-        </LoginWrapper >
+        </LoginWrapper>
     );
 };
